+++ .gitignore
@@ -20,3 +20,4 @@ hawaii
 2020B
 *.worksheet.sc
 marie
+marie-old
