+++ modules/engine/src/main/scala/api/QueueEngine.scala
@@ -9,5 +9,5 @@ import edu.gemini.tac.qengine.ctx.Partner
  * The queue generation engine.
  */
 trait QueueEngine {
-  def calc(proposals: List[Proposal], queueTime: QueueTime, config: QueueEngineConfig, partners: List[Partner]): QueueCalc
+  def calc(proposals: List[Proposal], queueTime: QueueTime, config: QueueEngineConfig, partners: List[Partner], extras: List[Proposal] = Nil): QueueCalc
 }
