+++ modules/main/src/main/scala/SummaryDebug.scala
@@ -7,6 +7,10 @@ import cats.implicits._
 import scala.collection.JavaConverters._
 import edu.gemini.model.p1.mutable._
 import gsp.math._
+import io.circe._
+import io.circe.syntax._
+import scala.math.BigDecimal.RoundingMode
+import io.circe.yaml.Printer
 
 /**
  * Utility to create a summary of the internal structure of a mutable P1 proposal, showing sharing.
@@ -14,68 +18,127 @@ import gsp.math._
  */
 object SummaryDebug {
 
-  def summary(pc: ProposalClass): String =
-    pc.getClass.getName
+  implicit def encoderEnum[A <: Enum[A]]: Encoder[A] = a =>
+    Json.fromString(a.name)
 
-  def summary(acc: SubmissionAccept): String =
-    if (acc == null) ""
-    else f"Accept:   ${acc.getRecommend.getValue().doubleValue()}%5.1f ${acc.getRecommend().getUnits()} at rank ${acc.getRanking()}, contact ${acc.getEmail()}"
+  implicit val EncoderProposalClass: Encoder[ProposalClass] =
+    Encoder[String].contramap(_.getClass.getName)
 
-  def summary(rej: SubmissionReject): String =
-    if (rej == null) "" else "Reject"
+  def summary(ta: TimeAmount): String =
+    s"${ta.getValue.setScale(1, RoundingMode.HALF_UP)} ${ta.getUnits}"
 
-  def summary(res: SubmissionResponse): String =
-    if (res == null) ""
-    else f"""|
-             |      Reference: ${res.getReceipt.getId}%-15s
-             |      ${summary(res.getAccept())}${summary(res.getReject())}
-             |      Comment:   ${res.getComment()}""".stripMargin
+  implicit val EncoderTimeAmount: Encoder[TimeAmount] = ta =>
+    summary(ta).asJson
+    // Json.obj(
+    //   "Value" -> ta.getValue.setScale(1, RoundingMode.HALF_UP).asJson,
+    //   "Units" -> ta.getUnits.asJson,
+    // )
 
-  def summary(sub: NgoSubmission): String =
-    s"""|- ${sub.getPartner()}:
-        |    Lead:    ${sub.getPartnerLead().getLastName()} ... ${System.identityHashCode(sub.getPartnerLead())}
-        |    Response:${summary(sub.getResponse)}""".stripMargin
+  implicit val EncoderSubmissionAccept: Encoder[SubmissionAccept] = acc =>
+    Json.obj(
+      "Recommend" -> acc.getRecommend.asJson,
+      "Email"     -> acc.getEmail.asJson,
+      "Ranking"   -> acc.getRanking.asJson,
+    )
 
-  def summary(rej: ItacReject): String =
-    if (rej == null) "" else "Reject"
+  implicit val EncoderSubmissionReject: Encoder[SubmissionReject] = _ =>
+    "Reject".asJson
 
-  def summary(acc: ItacAccept): String =
-    s"""|Accept:
-        |      Award:     ${acc.getAward().getValue()} ${acc.getAward().getUnits()}
-        |      Band:      ${acc.getBand()}
-        |      Contact:   ${acc.getContact()}
-        |      Email:     ${acc.getEmail()}
-        |      ProgramId: ${acc.getProgramId()}
-        |""".stripMargin.trim
+  implicit val EncoderSubmissionReceipt: Encoder[SubmissionReceipt] = rec =>
+    Json.obj(
+      "Contact" -> Option(rec.getContact).asJson,
+      "Id"      -> Option(rec.getId).asJson,
+    )
 
-  def summary(itac: Itac): String =
-    if (itac == null) "--"
-    else s"""|${summary(itac.getAccept())}${summary(itac.getReject())}
-             |    Comment: ${itac.getComment()}
-             |""".stripMargin.trim
+  implicit val EncoderSubmissionResponse: Encoder[SubmissionResponse] = res =>
+    Json.obj(
+      "Receipt" -> Option(res.getReceipt).asJson,
+      "Accept"  -> Option(res.getAccept).asJson,
+      "Reject"  -> Option(res.getReject).asJson,
+      "Comment" -> Option(res.getComment).asJson,
+    )
 
-  def summary(pc: QueueProposalClass): String =
-    s"""|  TooOption: ${pc.getTooOption()}
-        |  Exchange:  ${pc.getExchange()}
-        |  NGO:
-        |  ${pc.getNgo().asScala.toList.map(summary).mkString("\n  ")}
-        |  ITAC:
-        |    ${summary(pc.getItac())}
-        |""".stripMargin.trim
+  implicit val EncoderInvestigator: Encoder[Investigator] = inv =>
+    Json.fromString(s"${inv.getLastName} #<${System.identityHashCode(inv).toHexString}>")
+    // Json.obj(
+    //   "Email"     -> inv.getEmail.asJson,
+    //   "FirstName" -> inv.getFirstName.asJson,
+    //   "LastName"  -> inv.getLastName.asJson,
+    //   "Status"    -> inv.getStatus.asJson,
+    //   "HashCode"  -> System.identityHashCode(inv).asJson,
+    // )
 
-  def summary(pc: ProposalClassChoice): String = {
-    Option(pc.getClassical)     .map(summary) <+>
-    Option(pc.getExchange)      .map(summary) <+>
-    Option(pc.getFastTurnaround).map(summary) <+>
-    Option(pc.getLarge)         .map(summary) <+>
-    Option(pc.getQueue)         .map(summary) <+>
-    Option(pc.getSip)           .map(summary) <+>
-    Option(pc.getSpecial)       .map(summary)
-  } .get
+  implicit val EncoderPrincipalInvestigator: Encoder[PrincipalInvestigator] = inv =>
+    Json.obj(
+      "LastName" -> Json.fromString(s"${inv.getLastName} #<${System.identityHashCode(inv).toHexString}>"),
+      "InstitutionAddress" -> Json.obj(
+        "Institution" -> inv.getAddress().getInstitution.asJson,
+        "Country"     -> inv.getAddress().getCountry.asJson,
+      ),
+    )
 
-  def summary(c: Condition): String = {
-    f"- ${c.getId}%-12s ${c.getCc}%-5s ${c.getIq}%-5s ${c.getSb}%-5s ${c.getWv}%-5s"
-  }
+  implicit val EncoderSubmissionRequest: Encoder[SubmissionRequest] = _ =>
+    "<not implemented>".asJson
+
+  implicit val EncoderNgoSubmission: Encoder[NgoSubmission] = sub =>
+    Json.obj(
+      "Partner"     -> Option(sub.getPartner).asJson,
+      "PartnerLead" -> Option(sub.getPartnerLead).asJson,
+      // "Request"     -> Option(sub.getRequest).asJson,
+      "Response"    -> Option(sub.getResponse).asJson,
+    )
+
+  implicit val EncoderExchangeSubmission: Encoder[ExchangeSubmission] = sub =>
+    Json.obj(
+      "Partner"     -> Option(sub.getPartner).asJson,
+      "PartnerLead" -> Option(sub.getPartnerLead).asJson,
+      // "Request"     -> Option(sub.getRequest).asJson,
+      "Response"    -> Option(sub.getResponse).asJson,
+    )
+
+  implicit val EncoderItacReject: Encoder[ItacReject] = _ =>
+    "Reject".asJson
+
+  implicit val EncoderItacAccept: Encoder[ItacAccept] = acc =>
+    Json.obj(
+      "Award"            -> Option(acc.getAward).asJson,
+      "Band"             -> Option(acc.getBand).asJson,
+      "Contact (Gemini)" -> Option(acc.getContact).asJson,
+      "Email (NGO)"      -> Option(acc.getEmail).asJson,
+      "ProgramId"        -> Option(acc.getProgramId).asJson,
+    )
+
+  implicit val EncoderItac: Encoder[Itac] = itac =>
+    Json.obj(
+      "Accept"       -> Option(itac.getAccept).asJson,
+      "Reject"       -> Option(itac.getReject).asJson,
+      "Comment"      -> Option(itac.getComment).asJson,
+      "NgoAuthority" -> Option(itac.getNgoauthority).asJson,
+    )
+
+  implicit val EncoderQueueProposalClass: Encoder[QueueProposalClass] = pc =>
+    Json.obj(
+      // "Band3Request" -> Option(pc.getBand3Request).asJson,
+      "TooOption"    -> Option(pc.getTooOption).asJson,
+      // "Comment"      -> Option(pc.getComment).asJson,
+      "Exchange"     -> Option(pc.getExchange).asJson,
+      "Ngo"          -> Option(pc.getNgo).map(_.asScala).asJson,
+      "Itac"         -> Option(pc.getItac).asJson,
+    )
+
+  implicit val EncoderProposalClassChoice: Encoder[ProposalClassChoice] = pc =>
+    Json.obj(
+      "Classical"      -> Option(pc.getClassical).widen[ProposalClass].asJson,
+      "Exchange"       -> Option(pc.getExchange).widen[ProposalClass].asJson,
+      "FastTurnaround" -> Option(pc.getFastTurnaround).widen[ProposalClass].asJson,
+      "Large"          -> Option(pc.getLarge).widen[ProposalClass].asJson,
+      "Queue"          -> Option(pc.getQueue).asJson, // only specialized one right now
+      "Sip"            -> Option(pc.getSip).widen[ProposalClass].asJson,
+      "Special"        -> Option(pc.getSpecial).widen[ProposalClass].asJson,
+    )
+
+  implicit val EncoderCondition: Encoder[Condition] = c =>
+    f"${c.getId}%-15s ${c.getCc}%-5s ${c.getIq}%-5s ${c.getSb}%-5s ${c.getWv}%-5s".asJson
 
   def summary(d: DegDegCoordinates): String = {
     val ra  = HourAngle.HMS(Angle.hourAngle.get(Angle.fromDoubleDegrees(d.getRa.doubleValue))).format
@@ -83,39 +146,39 @@ object SummaryDebug {
     s"$ra $dec"
   }
 
-  def summary(t: Target): String = {
+  implicit val EncodeTarget: Encoder[Target] = t => {
     t match {
-      case t: SiderealTarget    => f"- ${t.getId}%-12s ${summary(t.getDegDeg)} ${t.getName}"
-      case t: NonSiderealTarget => f"- ${t.getId}%-12s <Non-Sidereal>                   ${t.getName}"
-      case t: TooTarget         => f"- ${t.getId}%-12s <ToO>                            ${t.getName}"
+      case t: SiderealTarget    => f"${t.getId}%-15s ${summary(t.getDegDeg)} ${t.getName}"
+      case t: NonSiderealTarget => f"${t.getId}%-15s <Non-Sidereal>                   ${t.getName}"
+      case t: TooTarget         => f"${t.getId}%-15s <ToO>                            ${t.getName}"
     }
-  }
+  } .asJson
 
-  def summary(o: Observation): String = {
-    f"* ${o.getBand}%-8s ${summary(o.getCondition)} ${summary(o.getTarget)}"
-  }
+  implicit val EncoderObservation: Encoder[Observation] = o =>
+    f"${o.getBand}%-9s ${o.getBlueprint().getId()}%-15s ${o.getCondition().getId()}%-15s ${o.getTarget().getId()}%-15s ${summary(o.getTime)}".asJson
 
-  def summary(b: BlueprintBase): String =
-    f"* ${b.getId}%-15s ${b.getName}"
+  implicit val EncoderBlueprintBase: Encoder[BlueprintBase] = b =>
+    f"${b.getId}%-15s ${b.getName}".asJson
+
+  implicit val EncoderInvestigators: Encoder[Investigators] = is =>
+    Json.obj(
+      "Pi"  -> is.getPi.asJson,
+      "Coi" -> Option(is.getCoi).map(_.asScala.toList.widen[Investigator]).asJson,
+    )
+
+  implicit val EncoderProposal: Encoder[Proposal] = p =>
+    Json.obj(
+      "Investigators" -> p.getInvestigators.asJson,
+      "Targets"       -> p.getTargets().getSiderealOrNonsiderealOrToo().asScala.asJson,
+      "Conditions"    -> p.getConditions().getCondition().asScala.asJson,
+      "Blueprints"    -> Merge.allBlueprints(p).asJson,
+      "Observations"  -> p.getObservations().getObservation().asScala.asJson,
+      "Scheduling"    -> Option(p.getScheduling().trim).filterNot(_.isEmpty).asJson,
+      "ProposalClass" -> p.getProposalClass.asJson,
+    )
 
   def summary(p: Proposal): String = {
-    f"""|PI: ${p.getInvestigators.getPi.getLastName()} ... ${System.identityHashCode(p.getInvestigators.getPi)}
-        |
-        |ProposalClass:
-        |  ${summary(p.getProposalClass)}
-        |
-        |Conditions:
-        |${p.getConditions.getCondition.asScala.map(summary).mkString("\n")}
-        |
-        |Targets:
-        |${p.getTargets.getSiderealOrNonsiderealOrToo.asScala.map(summary).mkString("\n")}
-        |
-        |Blueprints:
-        |${Merge.allBlueprints(p).map(summary).mkString("\n")}
-        |
-        |Observations:
-        |${p.getObservations.getObservation.asScala.map(summary).mkString("\n")}
-        |""".stripMargin
+    Printer(preserveOrder = true, dropNullKeys = true).pretty(p.asJson)
   }
 
 }
